fix: accuracy

+++ src/nn/demo/main.java
@@ -68,7 +68,7 @@ public static void main(String[] args){
     }   
 
     writer.close();
-    System.out.println("Test Accuracy: " + ((double)counter / 10000.0) * 100.0 + "%");
+    System.out.println("Test Accuracy: " + (counter / 10000) * 100.0 + "%");
 
     writer = new FileWriter("train_predictions.csv");
     Iterator<Vector> trainEval = trainSetEval.iterator();
@@ -85,7 +85,7 @@ public static void main(String[] args){
         }
     }
     writer.close();
-    System.out.println("Train Accuracy: " + ((double)counter / 60000.0) * 100.0 + "%");
+    System.out.println("Train Accuracy: " + (counter / 60000) * 100.0 + "%");
 
     long endTime = System.currentTimeMillis();
     System.out.println("Total time: " + (endTime - startTime)/1000.0 + " seconds");
